// src/types/pack.ts
export interface Pack {
  _id: string;
  id?: string;
  sous_categorie_id?: string | null;
  designation_fr?: string;
  cover?: string;
  qte?: string;
  prix_ht?: string | null;
  prix?: string;
  promo_ht?: string | null;
  promo?: string;
  description_fr?: string;
  publier?: string;
  created_by?: string | null;
  updated_by?: string | null;
  created_at?: string;
  updated_at?: string;
  code_product?: string | null;
  slug?: string;
  pack?: string;
  brand_id?: string | null;
  new_product?: string;
  best_seller?: string;
  gallery?: string;
  note?: string;
  meta_description_fr?: string;
  promo_expiration_date?: string | null;
  rupture?: string;
  alt_cover?: string;
  description_cover?: string;
  meta?: string;
  content_seo?: string;
  review?: string;
  aggregateRating?: string;
  nutrition_values?: string;
  questions?: string;
  zone1?: string;
  zone2?: string;
  zone3?: string;
  zone4?: string;
  mainImage?: {
    url: string;
    img_id: string;
  };
  [key: string]: any;
}